update rating type of difficulty and helpful from Integer to Float

+++ db/migrations/012_helpfuls_create.rb
@@ -5,7 +5,7 @@ Sequel.migration do
   change do
     create_table(:helpfuls) do
       primary_key :id
-      Integer :rating
+      Float :rating
     end
   end
 end
